Refactor Architecture
