+++ frontend/src/components/AuthButtons/index.ts
@@ -1,0 +1,2 @@
+export * from './AuthButtons';
+export * from './types.d';
