+++ frontend/src/components/AuthButtons/types.d.ts
@@ -1,0 +1,5 @@
+export type AuthButtonsProps = {
+  login: () => void;
+  logout: () => void;
+  isAuthenticated: boolean;
+};
