+++ frontend/src/components/Breadcrumbs/index.ts
@@ -1,1 +1,2 @@
 export * from './Breadcrumbs';
+export * from './types.d';
