+++ frontend/src/components/cards/PackageCard/index.ts
@@ -1,0 +1,2 @@
+export * from './PackageCard';
+export * from './types.d';
