+++ frontend/src/components/common/Card/types.d.ts
