+++ frontend/src/components/common/Heading/index.ts
@@ -1,0 +1,2 @@
+export * from './Heading';
+export * from './types.d';
