+++ frontend/src/components/common/Heading/types.d.ts
