+++ frontend/src/components/common/Modal/index.ts
@@ -1,0 +1,2 @@
+export * from './Modal';
+export * from './types.d';
