+++ frontend/src/components/common/PageLayout/index.ts
@@ -1,0 +1,2 @@
+export * from './PageLayout';
+export * from './types.d';
