+++ frontend/src/components/common/PageLayout/styles.ts
@@ -1,0 +1,4 @@
+export const styles = {
+  margin: '20px 14%', // Top/Bottom: 20px, Left/Right: 14% (responsive)
+  minHeight: 'calc(100dvh - 105px)', // Header: 65px + Margin: 40px = 105px offset
+};
