+++ frontend/src/components/common/PageLayout/types.d.ts
@@ -1,0 +1,3 @@
+export type PageLayoutProps = {
+  children: ReactNode;
+};
