+++ frontend/src/components/common/SectionCard/SectionCard.tsx
