+++ frontend/src/components/common/SectionCard/index.ts
@@ -1,0 +1,2 @@
+export * from './SectionCard';
+export * from './types.d';
