+++ frontend/src/components/common/SectionCard/types.d.ts
