+++ frontend/src/components/common/Spinner/index.ts
@@ -1,0 +1,2 @@
+export * from './Spinner';
+export * from './types.d';
