+++ frontend/src/components/common/Spinner/types.d.ts
