+++ frontend/src/components/common/Stack/Stack.tsx
