+++ frontend/src/components/common/Stack/index.ts
@@ -1,0 +1,2 @@
+export * from './Stack';
+export * from './types.d';
