+++ frontend/src/components/common/Stack/types.d.ts
