+++ frontend/src/components/common/Typography/index.ts
@@ -1,0 +1,2 @@
+export * from './Typography';
+export * from './types.d';
