+++ frontend/src/components/common/Typography/types.d.ts
