+++ frontend/src/components/common/index.ts
@@ -1,9 +1,9 @@
 export * from './Card';
-export * from './Stack';
 export * from './Heading';
-export * from './Typography';
 export * from './Icon';
-export * from './SectionCard';
-export * from './PageLayout';
-export * from './Spinner';
 export * from './Modal';
+export * from './PageLayout';
+export * from './SectionCard';
+export * from './Spinner';
+export * from './Stack';
+export * from './Typography';
