+++ frontend/src/components/github/GitHubIssues/index.ts
@@ -1,0 +1,2 @@
+export * from './GitHubIssues';
+export * from './types.d';
