+++ frontend/src/components/github/IssueCard/index.ts
@@ -1,0 +1,2 @@
+export * from './IssueCard';
+export * from './types.d';
