+++ frontend/src/components/github/IssueCard/types.d.ts
