+++ frontend/src/components/icons/Authenticate/Authenticate.tsx
