+++ frontend/src/components/icons/Authenticate/index.ts
@@ -1,0 +1,2 @@
+export * from './Authenticate';
+export * from './types.d';
