+++ frontend/src/components/icons/Authenticate/types.d.ts
