+++ frontend/src/components/icons/Exit/Exit.tsx
