+++ frontend/src/components/icons/Exit/index.ts
@@ -1,0 +1,2 @@
+export * from './Exit';
+export * from './types.d';
