+++ frontend/src/components/icons/Exit/types.d.ts
