+++ frontend/src/components/index.ts
@@ -1,3 +1,7 @@
+export * from './AuthButtons';
 export * from './Breadcrumbs';
+export * from './cards';
 export * from './common';
 export * from './github';
+export * from './icons';
+export * from './text';
