+++ frontend/src/components/text/Greeting/index.ts
@@ -1,0 +1,1 @@
+export * from './Greeting';
