+++ frontend/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useDynamicStyles';
+export * from './types.d';
