+++ frontend/src/hooks/types.d.ts
@@ -1,0 +1,1 @@
+export type StyleMapper<Props> = (props: Props, isHovered?: boolean) => CSSProperties;
