+++ frontend/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './endpoints';
 export * from './formatDateString';
-export * from './hasNoEmptyStringValues';
+export * from './formatTimeDifferenceString';
 export * from './getTimeBasedGreeting';
+export * from './hasNoEmptyStringValues';
